+++ include/optree/treespec.h
@@ -29,10 +29,10 @@ limitations under the License.
 
 #include <pybind11/pybind11.h>
 
-#include "include/exceptions.h"
-#include "include/hashing.h"
-#include "include/registry.h"
-#include "include/synchronization.h"
+#include "optree/exceptions.h"
+#include "optree/hashing.h"
+#include "optree/registry.h"
+#include "optree/synchronization.h"
 
 namespace optree {
