+++ src/treespec/flatten.cpp
@@ -24,12 +24,7 @@ limitations under the License.
 #include <utility>    // std::move, std::pair, std::make_pair
 #include <vector>     // std::vector
 
-#include "include/exceptions.h"
-#include "include/pytypes.h"
-#include "include/registry.h"
-#include "include/stdutils.h"
-#include "include/synchronization.h"
-#include "include/treespec.h"
+#include "optree/optree.h"
 
 namespace optree {
