+++ src/treespec/unflatten.cpp
@@ -18,11 +18,7 @@ limitations under the License.
 #include <sstream>  // std::ostringstream
 #include <utility>  // std::move
 
-#include "include/exceptions.h"
-#include "include/registry.h"
-#include "include/stdutils.h"
-#include "include/synchronization.h"
-#include "include/treespec.h"
+#include "optree/optree.h"
 
 namespace optree {
